chore: move plot-related utilities into plot folder
