+++ v3/src/components/graph/plots/bar-utils.ts
