+++ v3/src/components/graph/plots/dot-plot/dot-line-plot.tsx
@@ -3,15 +3,15 @@ import {observer} from "mobx-react-lite"
 import React, {useCallback, useEffect} from "react"
 import { mstAutorun } from "../../../../utilities/mst-autorun"
 import {mstReaction} from "../../../../utilities/mst-reaction"
+import { AxisPlace } from "../../../axis/axis-types"
 import { kMain } from "../../../data-display/data-display-types"
+import {circleAnchor, hBarAnchor, vBarAnchor} from "../../../data-display/pixi/pixi-points"
 import {IPlotProps} from "../../graphing-types"
+import { useDotPlot } from "../../hooks/use-dot-plot"
+import { useDotPlotDragDrop } from "../../hooks/use-dot-plot-drag-drop"
 import {usePixiDragHandlers, usePlotResponders} from "../../hooks/use-plot"
 import {setNiceDomain, setPointCoordinates} from "../../utilities/graph-utils"
-import {circleAnchor, hBarAnchor, vBarAnchor} from "../../../data-display/pixi/pixi-points"
-import { computeBinPlacements, computePrimaryCoord, computeSecondaryCoord } from "../../utilities/dot-plot-utils"
-import { useDotPlotDragDrop } from "../../hooks/use-dot-plot-drag-drop"
-import { AxisPlace } from "../../../axis/axis-types"
-import { useDotPlot } from "../../hooks/use-dot-plot"
+import { computeBinPlacements, computePrimaryCoord, computeSecondaryCoord } from "./dot-plot-utils"
 
 export const DotLinePlot = observer(function DotLinePlot({ pixiPoints }: IPlotProps) {
   const { dataset, dataConfig, graphModel, isAnimating, layout,
