+++ v3/src/components/graph/plots/dot-plot/dot-plot-utils.test.ts
